refactor: Split Prover and Verifier in separate files
# Conflicts:
#	wasm/prover/src/lib.rs
